requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Mail==0.9.1
Flask-CORS==4.0.0
Werkzeug==2.3.7
WTForms==3.0.1
psycopg2-binary==2.9.7
cryptography==41.0.7
Pillow==10.0.1
python-dotenv==1.0.0
bcrypt==4.0.1
itsdangerous==2.1.2
email-validator==2.0.0
gunicorn==21.2.0
requests==2.31.0
